requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# TakeHomeTask/venv/lib/python3.8/site-packages/starlette/templating.py: 8
# TakeHomeTask/venv/lib64/python3.8/site-packages/starlette/templating.py: 8
Jinja2 == 2.10.1

# TakeHomeTask/venv/lib/python3.8/site-packages/faker/providers/misc/__init__.py: 326,327
# TakeHomeTask/venv/lib64/python3.8/site-packages/faker/providers/misc/__init__.py: 326,327
Pillow == 7.0.0

# TakeHomeTask/venv/lib/python3.8/site-packages/httpcore/_backends/anyio.py: 7
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpcore/_backends/anyio.py: 7
anyio == 3.4.0

# TakeHomeTask/app.py: 4
# TakeHomeTask/backend.py: 2
boto3 == 1.20.26

# TakeHomeTask/venv/lib/python3.8/site-packages/click/_compat.py: 541
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/_compat.py: 541
colorama == 0.4.3

# TakeHomeTask/venv/lib/python3.8/site-packages/httpcore/_backends/curio.py: 4,5
# TakeHomeTask/venv/lib/python3.8/site-packages/httpx/_utils.py: 406
# TakeHomeTask/venv/lib/python3.8/site-packages/sniffio/_impl.py: 61
# TakeHomeTask/venv/lib/python3.8/site-packages/sniffio/_tests/test_sniffio.py: 50
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpcore/_backends/curio.py: 4,5
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpx/_utils.py: 406
# TakeHomeTask/venv/lib64/python3.8/site-packages/sniffio/_impl.py: 61
# TakeHomeTask/venv/lib64/python3.8/site-packages/sniffio/_tests/test_sniffio.py: 50
curio == 1.5

# TakeHomeTask/venv/lib/python3.8/site-packages/pkg_resources/_vendor/appdirs.py: 461
# TakeHomeTask/venv/lib64/python3.8/site-packages/pkg_resources/_vendor/appdirs.py: 461
future == 0.18.2

# TakeHomeTask/venv/lib/python3.8/site-packages/uvicorn/workers.py: 4
# TakeHomeTask/venv/lib64/python3.8/site-packages/uvicorn/workers.py: 4
gunicorn == 20.1.0

# TakeHomeTask/app.py: 2
# TakeHomeTask/dashboard.py: 2
# TakeHomeTask/new_user.py: 2
# TakeHomeTask/plot.py: 1
# TakeHomeTask/table.py: 1
h2o_wave == 0.19.0

# TakeHomeTask/venv/lib/python3.8/site-packages/click/decorators.py: 375
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/decorators.py: 375
importlib_metadata == 1.5.0

# TakeHomeTask/venv/lib/python3.8/site-packages/starlette/middleware/sessions.py: 6
# TakeHomeTask/venv/lib64/python3.8/site-packages/starlette/middleware/sessions.py: 6
itsdangerous == 2.0.1

# TakeHomeTask/venv/lib/python3.8/site-packages/pip/_internal/network/auth.py: 36
# TakeHomeTask/venv/lib64/python3.8/site-packages/pip/_internal/network/auth.py: 36
keyring == 18.0.1

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/sandbox.py: 18
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/sandbox.py: 18
mrlpy == 1.0.0

# TakeHomeTask/venv/lib/python3.8/site-packages/pip/_internal/operations/prepare.py: 50
# TakeHomeTask/venv/lib64/python3.8/site-packages/pip/_internal/operations/prepare.py: 50
mypy_extensions == 0.4.3

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/msvc.py: 184
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/msvc.py: 184
numpy == 1.21.4

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/sandbox.py: 18
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/sandbox.py: 18
openbaton-cli == 6.0.0

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/sandbox.py: 18
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/sandbox.py: 18
org-geppetto-recording == 0.1.0

# TakeHomeTask/app.py: 3
# TakeHomeTask/backend.py: 1
# TakeHomeTask/new_user.py: 3
pandas == 1.3.5

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/sandbox.py: 18
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/sandbox.py: 18
py4grid == 1.0.5

# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_connection.py: 1
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_events.py: 3
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_headers.py: 1
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_io.py: 1
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_receivebuffer.py: 3
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_state.py: 1
# TakeHomeTask/venv/lib/python3.8/site-packages/h11/tests/test_util.py: 5
# TakeHomeTask/venv/lib/python3.8/site-packages/sniffio/_tests/test_sniffio.py: 3
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_connection.py: 1
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_events.py: 3
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_headers.py: 1
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_io.py: 1
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_receivebuffer.py: 3
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_state.py: 1
# TakeHomeTask/venv/lib64/python3.8/site-packages/h11/tests/test_util.py: 5
# TakeHomeTask/venv/lib64/python3.8/site-packages/sniffio/_tests/test_sniffio.py: 3
pytest == 6.2.5

# TakeHomeTask/venv/lib/python3.8/site-packages/setuptools/sandbox.py: 18
# TakeHomeTask/venv/lib64/python3.8/site-packages/setuptools/sandbox.py: 18
python-plugin-sdk == 6.0.1

# TakeHomeTask/app.py: 5
# TakeHomeTask/backend.py: 3
# TakeHomeTask/venv/lib/python3.8/site-packages/uvicorn/config.py: 226
# TakeHomeTask/venv/lib64/python3.8/site-packages/uvicorn/config.py: 226
python_dotenv == 0.19.2

# TakeHomeTask/venv/lib/python3.8/site-packages/dateutil/rrule.py: 20
# TakeHomeTask/venv/lib/python3.8/site-packages/dateutil/tz/_factories.py: 5
# TakeHomeTask/venv/lib/python3.8/site-packages/dateutil/tz/tz.py: 21
# TakeHomeTask/venv/lib/python3.8/site-packages/dateutil/tz/win.py: 13
# TakeHomeTask/venv/lib64/python3.8/site-packages/dateutil/rrule.py: 20
# TakeHomeTask/venv/lib64/python3.8/site-packages/dateutil/tz/_factories.py: 5
# TakeHomeTask/venv/lib64/python3.8/site-packages/dateutil/tz/tz.py: 21
# TakeHomeTask/venv/lib64/python3.8/site-packages/dateutil/tz/win.py: 13
six == 1.14.0

# TakeHomeTask/venv/lib/python3.8/site-packages/httpcore/_backends/trio.py: 4
# TakeHomeTask/venv/lib/python3.8/site-packages/httpx/_transports/asgi.py: 10,17
# TakeHomeTask/venv/lib/python3.8/site-packages/httpx/_utils.py: 402
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpcore/_backends/trio.py: 4
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpx/_transports/asgi.py: 10,17
# TakeHomeTask/venv/lib64/python3.8/site-packages/httpx/_utils.py: 402
trio == 0.19.0

# TakeHomeTask/venv/lib/python3.8/site-packages/click/core.py: 42
# TakeHomeTask/venv/lib/python3.8/site-packages/click/globals.py: 6
# TakeHomeTask/venv/lib/python3.8/site-packages/click/parser.py: 35
# TakeHomeTask/venv/lib/python3.8/site-packages/click/types.py: 16
# TakeHomeTask/venv/lib/python3.8/site-packages/click/utils.py: 21
# TakeHomeTask/venv/lib/python3.8/site-packages/faker/typing.py: 7
# TakeHomeTask/venv/lib/python3.8/site-packages/uvicorn/_types.py: 5
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/core.py: 42
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/globals.py: 6
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/parser.py: 35
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/types.py: 16
# TakeHomeTask/venv/lib64/python3.8/site-packages/click/utils.py: 21
# TakeHomeTask/venv/lib64/python3.8/site-packages/faker/typing.py: 7
# TakeHomeTask/venv/lib64/python3.8/site-packages/uvicorn/_types.py: 5
typing_extensions == 4.0.1

# TakeHomeTask/venv/lib/python3.8/site-packages/uvicorn/loops/uvloop.py: 3
# TakeHomeTask/venv/lib64/python3.8/site-packages/uvicorn/loops/uvloop.py: 3
uvloop == 0.16.0

# TakeHomeTask/venv/lib/python3.8/site-packages/uvicorn/supervisors/watchgodreload.py: 5
# TakeHomeTask/venv/lib64/python3.8/site-packages/uvicorn/supervisors/watchgodreload.py: 5
watchgod == 0.7
